requirements.txt
# Core AI & Web Framework
google-generativeai>=0.3.0
gradio>=4.0.0

# Data Science & Visualization
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Jupyter/Notebook Support
nbformat>=5.9.0
ipykernel>=6.0.0

# Environment & Utilities
python-dotenv>=1.0.0
Pillow>=10.0.0

# Additional AI (if using google-genai)
google-genai>=0.1.0